Build a Markdown XBlock based on the HTML XBlock by OpenCraft
As we had a need for an XBlock that allows course authors to edit
content in Markdown, we came across the HTML XBlock from OpenCraft.
After considering/suggesting joining the efforts and *adding* that
functionality to the already excisting XBlock
(Link to the discussion about this:
https://discuss.openedx.org/t/about-the-html-xblock-and-possible-added-features/3215),
we chose to build an XBlock for Markdown editing only.

This commit is the inital setup for the Markdown XBlock. We use the
HTML XBlock as a base and replace the editors (raw HTML and text)
with one, Markdown editor.