Fix course export
The Markdown content by default would be exported as an attribute
in the <markdown> tag.
Implement add_xml_to_node() to add the content between <markdown></markdown>.
In addition, implement parse_xml() to accomodate the export changes
and make sure that the Markdown content will be added to the data field.