Add ability to specify safe_mode flavor
In Markdown2, safe_mode is actually a string, currently supporting the
values "replace" and "escape". Support for the Boolean value is only
maintained for compatibility with Markdown, a Boolean value of true is
equivalent to "replace".

Retain the default of "replace", but add the ability to also set
"escape". Also, interpret True as equivalent to 'replace', and False
or None to mean to disable safe mode.

Add test cases for all of the above.