Remove Python 3.5 from tox and GitLab CI pipeline
Apply the same logic as in 839abff269897c9a653ff3809ab1c6af30107213:
we're still able to test locally with Python 3.5 (with "tox -e py35",
using built-in tox virtualenv patterns[1]), but we really don't want
to block on changes that break only on that release anymore.

Also, update the README to say that 3.8 is the minimum Python version
required.

Reference:
[1] https://tox.readthedocs.io/en/latest/example/basic.html